requirements.txt
aiosqlite==0.22.1
alembic==1.18.4
APScheduler==3.11.2
httpx==0.28.1
nest-asyncio==1.6.0
pydantic==2.12.5
python-dotenv==1.1.0
python-telegram-bot==22.7
SQLAlchemy==2.0.49
tzlocal==5.3.1
